requirements_upbitMT.txt
requests>=2.28.0
pandas>=2.0.0
openpyxl>=3.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
pyupbit>=1.0.0
